Set command name when resolving
When a command is created the description is already a property of the command class. As a command can have many names (alias) it is not set in the command class itself.

However sometimes there can be login INSIDE the command class than depends on the name of the command at runtime. This fix allows the command name to be set when it is resolved so that internal logic can use that name.